Refactor DAT tests to make them work

+++ contracts/dxd/utils/DecentralizedAutonomousTrustUpgrade.sol
@@ -2,15 +2,11 @@ pragma solidity 0.5.17;
 
 import "../DecentralizedAutonomousTrust.sol";
 
-contract DecentralizedAutonomousTrustUpgrade
-  is DecentralizedAutonomousTrust
-{
-    
+contract DecentralizedAutonomousTrustUpgrade is DecentralizedAutonomousTrust {
     uint256 public newNumber;
 
-    function saveNumber(uint _newNumber) public {
-      require(control == msg.sender);
-      newNumber = _newNumber;
+    function saveNumber(uint256 _newNumber) public {
+        require(control == msg.sender);
+        newNumber = _newNumber;
     }
-  
 }
